CK-0003 - Migrated outdated (and soon outdated) code to new file
